pyproject.toml
[project]
name = "fastapi-todo-backend"
version = "0.1.0"
description = "FastAPI backend for Todo application with JWT authentication"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.16",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.6",
    "psycopg[binary]>=3.1.12",
    "alembic>=1.13.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "passlib>=1.7.4",
    "argon2-cffi>=21.3.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# Define the package structure using find directive to automatically discover packages
[tool.setuptools.packages.find]
where = ["."]
include = ["phase_2_web*"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
]